+++ javav2/example_code/bedrock-runtime/src/main/java/com/example/bedrockruntime/models/amazon/embeddings/G1InvokeModelScenarios.java
@@ -1,7 +1,7 @@
 // Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package com.example.bedrockruntime.models.amazon.embeddings.text;
+package com.example.bedrockruntime.models.amazon.embeddings;
 
 import com.example.bedrockruntime.libs.ScenarioRunner;
 import org.json.JSONObject;
